MAINT: clean up, comment, split up analysis_widgets.py; ENH: better generic plotting, saving error traceback on subthreads to file
